+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 20.11.11
-* Hotfix for User profile data.
+* User profile issue fixed, user data was deleted for key if no value,null or undefined value was provided against that key.
 * Updated underlying android SDK to 20.11.11
 * Updated underlying iOS SDK to 20.11.3